+++ tpl/js/validator.js
@@ -1,6 +1,15 @@
+/**
+ * Sample {{type}} script.
+ *
+ * @author {{author}}
+ * @created {{date}}
+ */
 ;(function(app){
 
 	app.editor.validator('{{{name}}}', function(options, val, form){
+
+		//check the val here, or ref other fields in the form
+		//options mirrors what you put in the editor's configure block
 
 		//return nothing if ok
 		//return error message 'string' if not