+++ tpl/js/view.js
@@ -1,12 +1,26 @@
-//same as regional/area
+/**
+ * Sample {{type}} script. (named view, alias: Regional/Area)
+ *
+ * @author {{author}}
+ * @created {{date}}
+ */
 ;(function(app){
 
 	app.view({
 
-		name: '{{{name}}}',
+		name: '{{{title}}}',
 		template: '@{{{template}}}',
-		//...
 		
+		//it is basically a Backbone.View or Marionette.ItemView/Layout/CollectionView/CompositeView definition
+		
+		//initialize: function(){ ... },
+		//onShow: function(){ ... },
+		
+		//actions: {
+		//	submit: function(){...},
+		//	dosomething: function(){...},
+		//	...
+		//},
 	});
 
 })(Application);